feat: relocate file locking logic from main to root command and remove unused SearchQuery struct
